bedrock-kb-sync/requirements.txt
aws-cdk-lib>=2.195.0
constructs>=10.0.0,<11.0.0
boto3>=1.28.0
